+++ ibft/pipeline/preprepare/validate_test.go
@@ -112,7 +112,7 @@ func TestValidatePrePrepareValue(t *testing.T) {
 	}
 	for _, test := range tests {
 		t.Run(test.name, func(t *testing.T) {
-			err := ValidatePrePrepareMsg(consensus, func(round uint64) uint64 {
+			err := ValidatePrePrepareMsg(consensus, []byte{}, func(round uint64) uint64 {
 				return 1
 			}).Run(test.msg)
 			if len(test.err) > 0 {
