+++ utils/dataval/bytesval/equal_bytes.go
@@ -20,7 +20,7 @@ func NewEqualBytes(val []byte) valcheck.ValueCheck {
 }
 
 // Validate implements dataval.ValidatorStorage interface
-func (c *bytesValidation) Check(value []byte) error {
+func (c *bytesValidation) Check(value []byte, pk []byte) error {
 	if !bytes.Equal(value, c.val) {
 		return errors.New("msg value is wrong")
 	}
